requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.0.1
build==1.2.2
CacheControl==0.14.0
certifi==2024.8.30
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
dataclasses-json==0.6.7
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.2.0
Faker==29.0.0
fastapi==0.115.0
fastapi-cli==0.0.5
fastjsonschema==2.20.0
filelock==3.16.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
gritql==0.1.5
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==24.3.1
langchain-cli==0.0.31
langchain-community==0.0.38
langchain-core==0.3.5
langchain-groq==0.2.0
langserve==0.3.0
langsmith==0.1.128
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.47.1
orjson==3.10.7
packaging==23.2
passlib==1.7.4
pexpect==4.9.0
# Editable install with no version control (pirate-speak==0.0.1)
-e c:\users\penhs\seyha\fastapi\langserve-2\my-app\packages\pirate-speak
pkginfo==1.11.1
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg2==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sse-starlette==1.8.2
starlette==0.38.6
tenacity==8.5.0
tomlkit==0.12.5
tqdm==4.66.5
trove-classifiers==2024.9.12
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.23.2
virtualenv==20.26.5
watchfiles==0.24.0
websockets==13.1
yarl==1.12.1
